lib: Properly assign hostnames
The zookeeper spec relied on a correspondence between the order of
hostnames returned by `Service.children`, and the order the QRI assigned
hostnames to the containers within a service. This assumption is no
longer safe, so this patch migrates us to the new `setHostname`
function, which allows us to set the hostname of a container at
compile time.